+++ hrk_nasa_apis/lib/src/services/ssd_cneos/sbdb_cad_api/data/sbdb_cad_data.dart
@@ -61,5 +61,5 @@ class DistanceConverter implements JsonConverter<Distance, String> {
   }
 
   @override
-  String toJson(Distance distanceObject) => distanceObject.value!.toString();
+  String toJson(Distance distanceObject) => distanceObject.value.toString();
 }