+++ hrk_nasa_apis/test/services/ssd_cneos/sbdb_cad_api/data/sbdb_cad_query_parameters_test.dart
@@ -116,36 +116,16 @@ void main() {
         (unit) => unit != defaultUnit,
       );
 
-      test('null value and unit', () {
-        const distRange = DistanceRange();
-        queryParameters = queryParameters.copyWithDistRange(distRange);
+      test('null min max', () {
+        queryParameters = queryParameters.copyWithDistanceRange(null, null);
         expect(queryParameters.distMin, null);
         expect(queryParameters.distMax, null);
-      });
-
-      test('null value', () {
-        final distMin = Distance(unit: nonDefaultUnit);
-        final distMax = Distance(unit: nonDefaultUnit);
-        final distRange = DistanceRange(start: distMin, end: distMax);
-        queryParameters = queryParameters.copyWithDistRange(distRange);
-        expect(queryParameters.distMin, null);
-        expect(queryParameters.distMax, null);
-      });
-
-      test('null unit', () {
-        const distMin = Distance(value: 1);
-        const distMax = Distance(value: 2);
-        const distRange = DistanceRange(start: distMin, end: distMax);
-        queryParameters = queryParameters.copyWithDistRange(distRange);
-        expect(queryParameters.distMin, distMin.value.toString());
-        expect(queryParameters.distMax, distMax.value.toString());
       });
 
       test('defaultUnit', () {
         const distMin = Distance(value: 1, unit: defaultUnit);
         const distMax = Distance(value: 2, unit: defaultUnit);
-        const distRange = DistanceRange(start: distMin, end: distMax);
-        queryParameters = queryParameters.copyWithDistRange(distRange);
+        queryParameters = queryParameters.copyWithDistanceRange(distMin, distMax);
         expect(queryParameters.distMin, distMin.value.toString());
         expect(queryParameters.distMax, distMax.value.toString());
       });
@@ -153,8 +133,7 @@ void main() {
       test('nonDefaultUnit', () {
         final distMin = Distance(value: 1, unit: nonDefaultUnit);
         final distMax = Distance(value: 2, unit: nonDefaultUnit);
-        final distRange = DistanceRange(start: distMin, end: distMax);
-        queryParameters = queryParameters.copyWithDistRange(distRange);
+        queryParameters = queryParameters.copyWithDistanceRange(distMin, distMax);
         expect(queryParameters.distMin, distMin.toCompactString());
         expect(queryParameters.distMax, distMax.toCompactString());
       });